extract cover upload out into mixin
