+++ app/frontend/app/templates/components/group-box.hbs
@@ -1,7 +1,7 @@
 <div class="group-wrapper" {{bind-attr style=group.coverImageStyle}}>
   <div class="group-widget-top">
     <div class="group-avatar">
-      <img {{bind-attr src=group.avatar}}>
+      <img {{bind-attr src=group.avatarUrl}}>
     </div>
     <div class="group-info">
       <h4>{{#link-to 'group.index' group}}{{group.name}}{{/link-to}}</h4>
