+++ app/frontend/app/templates/group.hbs
@@ -27,7 +27,7 @@
         {{! Group Avatar }}
         <div class="col-sm-2 col-md-2 user-avatar clearfix">
           <div class="large-avatar">
-            <img class="responsive-image" {{bind-attr src=model.avatar alt=model.name}}>
+            <img class="responsive-image" {{bind-attr src=model.avatarUrl alt=model.name}}>
             <div class="edit-avatar">
               <label for="avatar-image">
                 <i class="fa fa-camera"></i> Edit Avatar
